% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localconf.R
\name{encrypt_conf}
\alias{encrypt_conf}
\title{Encrypt configuration details in project root}
\usage{
encrypt_conf(..., file_name = ".conf.rds")
}
\arguments{
\item{file_name}{}
}
\description{
Encrypt configuration details in project root
}
