% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localconf.R
\name{register_conf}
\alias{register_conf}
\title{Register configuration details as environment variables}
\usage{
register_conf(file_name = ".conf.rds", verbose = TRUE)
}
\arguments{
\item{file_name}{}
}
\description{
Register configuration details as environment variables
}
