% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localconf.R
\name{retrieve_conf}
\alias{retrieve_conf}
\title{Retrieve configuration setting from the environment}
\usage{
retrieve_conf(var_name)
}
\arguments{
\item{var_name}{}
}
\description{
Retrieve configuration setting from the environment
}
